Hide buttons on non-default layouts
